extract get_camera into own file
