+++ uap_tracker/main.py
@@ -16,6 +16,7 @@ from uap_tracker.camera_stream_controller import CameraStreamController
 from uap_tracker.tracker_listener_stf import TrackerListenerMOTStf, TrackerListenerSOTStf
 from config import settings
 from uap_tracker.video_tracker import VideoTracker
+from camera import get_camera
 
 USAGE = 'python uap_tracker/stage1.py\n settings are handled in the setttings.toml file or overridden in the ENV'
 
@@ -118,7 +119,7 @@ def process_file(controller, full_path):
 
     print(f"Opening {full_path}")
     video = cv2.VideoCapture(full_path)
-            # Exit if video not opened.
+    # Exit if video not opened.
     if not video.isOpened():
         print("Could not open video")
         sys.exit()
@@ -129,43 +130,6 @@ def process_file(controller, full_path):
         events.listen(listener)
 
     _run(controller, listener, video)
-
-
-def get_camera(config):
-    camera_mode = config.get('mode', 'rtsp')
-    camera_uri = config.get('camera_uri',
-                            'rtsp://admin:sky360@192.168.1.108:554/live')
-    print(f"Connecting to {camera_mode} camera at {camera_uri}")
-    camera = None
-    if camera_mode == 'ffmpeg':
-        os.environ["OPENCV_FFMPEG_CAPTURE_OPTIONS"] = "rtsp_transport;udp"
-        camera = cv2.VideoCapture(
-            camera_uri,
-            cv2.CAP_FFMPEG
-        )
-    elif camera_mode == 'rtsp':
-        camera = cv2.VideoCapture(
-            camera_uri
-        )
-    elif camera_mode == 'local':
-        for i in range(-1, 100):
-            try:
-                camera = cv2.VideoCapture(i)
-                if camera:
-                    break
-            except cv2.error as e:
-                print(e)
-            except Exception as e:
-                print(e)
-    ##
-    # Apparently this works for firewire, but I couldn't get it to
-    ## camera = cv2.VideoCapture(3, cv2.CAP_DC1394)
-
-    if not camera:
-        print(cv2.getBuildInformation())
-        print(f"Unable to find camera using config: {config}")
-        sys.exit(1)
-    return camera
 
 
 def _run(controller, listener, media):
